(#135) Resolving CVE-2020-26160.
Migrated from an unmaintained JWT library to one with a security patch
applied to it. This is an immediate fix to a problem. The library that
is now being used is a fork. In the future we might want to consider
moving to a library that we know for certain will be actively maintained
long term.