Improvements to configuration and progress on billing.
Billing features are now string arrays. This will make it easier to
understand them but also make it easier to create unique feature tags
for subscriptions.

Log level is now parsed from the configuration for the primary logger.
Other loggers default to FATAL level for now though to suppress noise.

Stopped using Trace() for logging except for database queries, the log
was really chatty with the queries, but there were useful messages that
were still a bit verbose. Changed all non-db-query log messages to be
debug instead of trace.

Billing stuff will now be primarily stored in the configuration. This
configuration will not be the end all be all for pricing. Long run I'm
still not sure how I want to handle pricing/billing. But storing the
basic identifying data in the config makes it easy right now. I will be
adding something that will periodically sync data from Stripe into the
database based off of those config entries. This way active pricing and
products are kept in the config and kept up to date automatically. And
as prices/products become out of date they are removed from the config.
Only prices/products that are in the config are ever shown in the UI as
selectable to the user. And only if they are set as visible.