Remember the user's selected bank account ID.
Store the selected bank account Id in local storage. This way the Id is
persisted accross sessions. If the Id is ever considered "invalid" for
the current list of bank account Ids, then remove it from local storage.

If there is no bank account selected, and there is nothing in local
storage, default to the first bank in our list of bank accounts.