Adding handling of removing links.
This adds support for removing manual and plaid links. Plaid links will
revoke their access token when the removal is initiated. But the actual
data removal for links is performed in the background asynchronously.
This is because of the potential amount of data being removed by this
job. I don't know how much data might need to be removed right now, but
I know it could be a lot someday.