fix: Stripe race condition.
This fixes a Stripe race condition that could happen after a user
initially subscribes to the application. This is because I was
previously relying on webhooks to determine the state of the
subscription. In testing these webhooks had always been delivered before
the user had returned to the monetr app. But I've since been able to
reproduce situations where the user would be redirected back to the
application before a webhook had been delivered. This resulted in the
application beliving the user did not yet have an active subscription
and directing them back to checkout. Potentially allowing the user to
create multiple subscriptions, further corrupting the billing state of
their account with monetr.

This resolves that issue by immediately checking the state of the
subscription that was created from the checkout session as soon as the
user returns to monetr from their stripe checkout. With this change in
approach I've also removed the `/billing/wait` endpoint as it does not
really make sense any more as the subscription is not necessarily
managed "asynchronously" the first time around anymore.