refactor: Build/push container as part of release.
This adds building of the container image for multiple (linux)
architectures when a release is performed. This particular commit is a
test and does push the container without a tag.

This also adds the version label to the built container.

TODO: Right now the BUILD_TIME variable is not included in the container
build during github actions.