fix: Fixed issue where the UI file names changed each build.
This fixes the issue where the UI file names would change during a
rollout deployment. Because the files are hosted inside the application
itself it would be possible for one version of the application to serve
the index.html that would tell the client to request static files that
might not be servable from all instances of the application. This
resolves the issue by serving the same file names always.

This will introduce some caching goofyness, but that should be easier to
handle overall.

Resolves #94