feat: Preventing duplicate item adds.
This adds Plaid's institution ID to the link object in the database and
in the API. It removed the previously unused institutions table (as it
will be replaced by the institution details API for a specific link).

This also adds a UI check that looks for an intersection between the new
link's `institution_id` and the bank account masks that the user
selected.

This allows us to prevent (at least on a basic level) a user accidently
adding a duplcate account. However; this does not outright prevent it
and it does still allow the user to override it and add the item anyway.

Resolves: #193