fix: Allow log format to be configured.
This adds support for setting the log format to either `text` or `json`
which will change the logrus formatter used. Invalid log options
provided will default to `text`.

This also makes a small tweak to how the release name is reported to
sentry via the SDK. This more closely aligns with their documentation,
but I'm not sure if it will have any affect on anything.

Resolves #237