fix: Added better Sentry error reporting.
This adds a dumb approach to a problem; I want to be able to configure
my Sentry at runtime, not compiletime. So when the application initially
loads it will request a public Sentry DSN from the API. Then once it has
that it will setup Sentry, then it will call ReactDOM.render in a
`finally` of the promise. This is super hacky, but should be just fine
for configuring this at runtime. Another option might be to add a flag
  in the UI to allow this behavior to be configured or outright disabled
  at compiletime.