fix: Adding security headers and removing stripe-js.
This adds security headers to the UI, and to reduce the amount of URLs
in our allow-list I have removed the stripe-js library from our UI as
well. Instead the API will generate the URL for the customer to be
redirected to. This way we don't need the JS library at all.

Resolves #304