feature: Switching to buildah for container builds.
Was having problems with Docker literally just not working, the same
Dockerfile could be built locally by docker, or podman. But in GitHub
Actions it would not build for some reason. The error was a go error,
but could not reproduce.

Docker is also extremly slow, so moving to buildah for container images
from now on.