fix: Fix page crash on changing bank account.
When changing bank accounts, if the balance for the newly selected bank
account is not loaded before the spent from selection component is
loaded then the page will crash.

This resolves that issue by defaulting to a string of "N/A" when the
safe to spend balance is not yet loaded for the transaction spent from
selection.

Resolves #700