feat(icons): Adding basic icon support.
Adding support for building out icons in the future. At the moment this
adds basic support for embedding "simple-icons" into the monetr
application using a build flag. Icons can be retrieved via an API call
that will perform a basic search using the provided string. In the
future I will improve this search, as well as how the data is indexed.