fix(job): Improving debugging around removing transactions.
There was an issue where some pending transactions were not properly
removed by the job. They seem to not have been in the database at all
which is bizarre. This adds some utility work around making situtions
like this more discoverable as well as include more information about
the arguments for the job when it happens.