fix(forecast): Reduce number of forecast requests, reduce sentry trace.
We were sending forecast requests every time a text input changed once
the form had been filled out. So adjustments in an expense would result
in odd behaviors.

I've also removed a lot of the sentry tracing from the forecasting code,
it was creating hundreds of spans because of how the code executes.