feat(api): Adding basic soft-delete for transactions.
This makes it so that transactions are no longer removed by the
background sync job. This will make it a bit easier to reconcile when a
transaction went from pending to a cleared state. But more work still
needs to be done to make the spendingId or other details persist between
transactions.

Related to #104