feat(ui): Improve offline state of the UI.
When the API not working properly the UI will now respond a bit better.
It's not a perfect solution because the UI files are usually served by
the same service that serves the API. However if the UI files are cached
and the server is generally just offline; this will be more graceful
with this change.