fix(api): Fixed expense being created with wrong first due date.
We were always calculating the next due date for expenses when they were
created; this caused us to have a mismatch with what the user input
sometimes.

This just makes it so as long as the provided date is not in the past we
accept the user provided next recurrence for the initial input.

Resolves #1170