fix(ui): Fix OAuth links not working if they were not the first.
If you setup a Plaid link and then attempt to add an additional link
that requires OAuth; it will fail because the OAuth return route was not
configured in the routes once a link exists.

This just adds the OAuth return route to to the route config when there
are links setup. This would also fix if you initially started with a
manual link and then tried to add an OAuth institution.

Resolves #488