feat(api): Add DTSTART equivalent support to spending/funding.
Right now we do not store the DTSTART component of the RRule, this has
led to some bugs already where we have to assume the DTSTART based on
some educated guess data. This can lead to wrong calculations in the
future. So going forward I want to support this parameter. This is the
start of the work on that.