fix(ui): Fixed requesting more transactions. (#1367)
Previously if you had a number of transactions exactly divisible by 25
we would continue to request more and more transactions unnecessarily.

This makes it so that we properly handle this by looking to see if we
have any pages with less than 25, which would include a page that is
empty.

Resolves #1331