feat(ui): Adding storybook support, improved webpack. (#1391)
This is a significant addition to the frontend development workflow. I wanted to be able to build out UI components and test them in various scenarios without needing to restart my development environment with new configurations.

This adds storybook support and a few new components. Long term I want to standardize how components are used in monetr and what components are used. Right now everything is a mashup of tailwind, react-select, and material UI.

Unfortunately I think material-ui is here to stay, however for many of the other components I need. I do not feel the need to use material ui going forward. Instead I want to try to write the most basic and maintainable components using tailwind. This will hopefully make iterating on those components easier, as well as theming them and implementing them throughout the UI.

---

This also includes a massive cleanup of the terrible lock file that was causing a nearly 4GB node_modules directory when running locally. And improves SVG and other image load patterns in the UI by embedding them in the actual bundle if they are under 4MB in size.