feat(ui): Building out new forgot password screen. (#1392)
This is the new forgot password screen wireframe. It isn't connected to
the actual application yet. But will be soon.

This also includes improvements to all of the form elements that have
been created. Automatically wiring these components up to Formik if the
context exists. This will reduce the number of parameters that need to
be passed to every input when building out forms for monetr.