feat(build): Build third party notice file using container.
Up to this we were using an outdated version of licensed that was
executed on the host machine, and behaved differently in CI. This is a
problem when trying to have consistent builds.

This moves licensed into a container file and builds the file at
buildtime instead of storing the file in git.