fix(plaid): Fixed an issue with Plaid update callback syncing.
When we would initiate a sync after a Plaid update callback had been
completed, we were always triggering a pull transactions sync. This is
the legacy method for syncing data, and the preferred method is the
regular plaid sync.

This resolve the issue by looking at the setting on the plaid link
itself to determine which type of sync to initiate.

Resolves #1445