fix(dev): Fixed permission issues on linux for local dev.
air would create a `tmp` dir in the project directory inside docker.
This directory would have root permissions which would cause issues when
someone wanted to run `make clean` afterwards. The directory couldn't be
removed without sudo.

This fixes that by making sure the directory has the right permissions
inside the container so that it can be removed outside if needed.