feat(api)!: Updating a funding schedule returns updated spending.
When a funding schedule is updated such that it's funding dates would
change. All of the spending objects associated with that funding
schedule are also now updated. This is a breaking change because the API
for updating a funding schedule now returns both the funding schedule
that was updated, as well as the spending objects that were updated.