feat(ui): Improved MSelectFrequency significantly.
MSelectFrequency allows the current frequency to be read now upon mount.
This means that for an edit modal or view, we can pre-select the current
rule based on the date context. Rules are now returned as strings too,
instead of returning a recurrence object.

This is also probably the beginning of removing the recurrence object
because we need to start incorporating date started.