feat(billing)!: monetr now tracks trial status instead of stripe.
Instead of having Stripe maintain the trial status, which is funky;
monetr will keep track of the trial status for an account. Then at the
end of the trial the user will be prompted with a Stripe checkout, which
will allow us to collect the information necessary to properly charge
tax (which is included in the price).