fix(ui): Fixed mobile menu not displaying properly on iOS.
I had at some point removed autoprefixer apparently, which auto fills a
ton of CSS stuff for various browsers. This caused the compiled CSS file
to not have anything for webkit, including blur stuff. So the browser on
mobile was just showing the menu without any of the fancier styling.

Resolves #1513