fix(ui): Fixed amount formatting in the UI.
This fixes the funding schedule details page amount formatting, as well
as every other place where amounts might be formatted. This makes it so
that every time a number is formatted it is done so using a common
function. This will also help support other currency codes in the
future.

Resolves #1509