fix(ui): Fix funding schedule recurrence select. (#1524)
This fixes the bug where if a funding schedule excludes weekends and has the nextOccurrence adjusted becaue of that;
the recurrence rule will no longer align with that date. This fixes it by making the next occurrence date picker only
pick the non-adjusted dates. So the date picker would show the original date that would have falled on a weekend.

This might be confusing but I think should be a more accurate depiction of how the date should be modified anyway?

It also adds some tests for the details page to proove that a label decorator appears when appropriate. The label
decorator will appear when there is that adjustment so that hopefully there is less confusion?

Resolves #1510