fix(api): Migrating to rruleset format for recurrence rules. (#1532)
This migrates all of the rrule instances to rrulesets, which are better because they include the date started in the rule itself. This requires some tweaks to a ton of tests and a kind of gnarly schema migration. But this will setup monetr to be a lot more consistent in the future about how spending is handled and calculated.

The dtstart is still stored in UTC though with the hours of the offset for that start time. This is corrected in the forecasting code where the dtstart is always adjusted to be the current timezone if possible. This means that crossing daylight savings will be correct for the rules, something that was iffy before. It also means that if TZData is not available for a rule, we can still get a good estimate of the timestamp within an hour or so.