feat(local): Updated docker compose for self-hosting. (#1545)
* feat(local): Updated docker compose for self-hosting.

The docker compose file in the project root directory now serves as a
very basic template for how monetr can be self hosted. It provides some
basic environment variable templates which can be provided in order to
mess around with it locally more easily.

This file should not be used as "the" way to host monetr locally.
Ideally you would not be building the container from scratch each time
and can instead simply pull the container from docker hub or ghcr. But
it is a good starting place.

Resolves #1536

* chore: Adding docs