fix(ui): Show loading state properly on transaction spending.
This would show the loading state on initial load even if it knew the
transaction did not have any spending assocaited with it. Not it will
only show the loading state when the selection is being changed. Or when
it is properly waiting for a spending object to be retrieved via the
API.