feat(build): Migrate build system to CMake. (#1495)
This migrates the build system that monetr uses for local development and CI/CD from vanilla GNU Make to CMake.

CMake allows monetr to more easily support other OSs besides macOS and Linux for local development by offering a scripting language that is agnostic to the host system that is running it. It also supports the more complex build workflows that monetr requires in order to produce a single monolithic binary. 

CMake is more complex than GNUMake, but offers a lot more power and customizability for complex workflows. I'm hoping that the migration to CMake will significantly improve monetr's local development experience by having the build tasks be consistent across operating systems as well as supporting a more diverse array of tooling on the host.