fix(forecast): Fixed bug with forecasting completed goals
Fix an issue where the forecasting code would not recognize that the
goal is complete and would assume that contributions needed to be made
to it equal to the `targetAmount - currentAmount`. This isn't correct
because goals have a used amount that keeps track of funds spent from
the goal. This makes it so the used amount is now taken into account for
contributions predicted.

Resolves #1561