fix(ui): I hate javascript so much
This fixes being able to run tests with shit as a sub project as part of
a workspace. This took so much time and ended up basically just being a
package upgrade. There were no issues on msw or swc or jest about this
from what i can tell. I hate javascript tooling so much