feat(ui): Moving user interface to workspace. (#1586)
* feat(ui): Moving user interface to workspace.

* chore: Fixing build

* chore: Fixing some stuff, tests broken

* chore: Fixing email things

* chore: What

* fix(ui): Fixed HMR for interface

* chore: Fix overrides

* fix(ui): I hate javascript so much

This fixes being able to run tests with shit as a sub project as part of
a workspace. This took so much time and ended up basically just being a
package upgrade. There were no issues on msw or swc or jest about this
from what i can tell. I hate javascript tooling so much

* chore: Tweak ctest paths for user interface

* chore: Clean up packages

* chore: `pkg` -> `server`

* fix: Licensed config

* chore: Tweak email dev

* chore: Tweak caching

* chore: Tweak caching again

* chore: Tweak air

* chore: Tweaks