feat(docs): Migrating docs to next.js (#1593)
This lays the ground work for moving the entire monetr.app away from mkdocs and over to nextra which is powered by Next.js. This will hopefully make it easier for anyone to contribute to monetr's documentation as it now only requires a node.js runtime in order to develop locally.