feat(api): Adding clock and moving to PASETO instead of JWT. (#1618)
This adds a clock system throughout monetr, allowing time to be mocked in tests to allow a greater array of functionality testing.

As part of this, the JWT library that was being used would not (in a thread safe way) allow time to be mocked. As such I have migrated monetr away from that library in favor of a PASETO library. 

This lets me consolidate all of the tokens needed in the application into a single token with a single public and private key. Tokens can then be used for a specific scope or reason by various parts of the application such as resetting passwords, verifying email addresses etc.