feat(api): Trying to find a way to isolate specific amounts.
Recurring transactions are complicated, I want to try to isolate
specific amounts within a dataset of known similar transactions. This
way I can determine which transactions are most likely to be recurring,
but I want to narrow this down to be more accurate. Some similar
transactions might actually be two subscriptions. Or there may be other
patterns. This is really just throwing stuff at the wall and seeing what
sticks