feat(api): Laying groundwork for storing recurring transaction data.
This doesn't include the DB migrations yet. I'll work on that shortly.
But this is basically what the model is going to look like and at least
for MVP it'll be the same as the similar transaction calculations. But
I'll need to make some tweaks to consume all of the similar transaction
clusters and then use those to produce recurring transaction clusters.
There are certainly a lot of optimizations that can be made here but
right now I just want to get this into a state where it can be used in
the UI.