fix(ui): Fixed mobile transaction header too tall on non-manual links.
This was caused by the upload button "maybe" rendering but causing the
styles to push the transaction list itself down from the header even
when it did not render. This resolves the problem by tweaking the styles
so that the list is only pushed down when a non null component is
rendered there.