fix(api): Fixed SMTP message formatting.
Because the HTML content was a single line it would get clobbered by
either sendgrid or something else when it was sent because it had lines
exceeding the 1000 character (or byte?) limit. This was causing the
verify email link to get fucked up. This adds something where itll
automatically cut the HTML content into lines 900 characters in length.
This seems to fix the problem locally.