feat(api): Rewriting how secrets are stored.
Right now secrets were stored in a plaid tokens table and I was going to
do something similar for teller. But it wasn't a great pattern and
caused some weirdness with dependency cycles. Instead I'm going to store
secrets in a separate table entirely. This table will store all secrets
instead of just secrets of a certain type. And secrets will be scoped to
a single account.