fix(api): Fix plain text secret encoding.
Plaintext secrets were meant to be stored as exclusively plain text. But
the changes to how secrets were stored is causing them to be
read/written as hexadecimal which is incorrect. This lives the
hexadecimal encoding out of the secrets repository and instead into the
key management implementation itself. If the implementation is actually
encrypting the secrets then they will be returned in a hexadecimal
format. Plaintext secrets will be stored unaltered.

Resolves #1673