fix(forecast): Fixed bug causing inconsistent contributions to goals.
When a funding schedule "excluded weekends" it's contributions to goals
over long periods of times could be very inconsistent which makes
planning harder to predict. The goal of the forecasting system is to
have consistent contributions throughout the lifetime of a goal or
expense.

This resolves the problem by making the exclude weekend code actually
work in one of the funding schedule interface functions, as well as
making it start exclusive and end inclusive in the spending query.