feat(api): Moving contribution calculation out of models
Right now monetr calculates contributions to expenses and goals using
some specific code co-located in the models package. This isn't great as
that package should not contain business logic as much as possible. This
is the groundwork of moving that logic out of that package and to make
use of the existing forecasting code to facilitate the same
calculations. This should also make things more consistent as both the
contribution and spending calculations will be the same code.