fix(background): Fixed timeouts in background tests
What was happening was that all of the background threads (1 cron, 1
consumer and 4 workers) were in a single array of channels for shutdown
signals. But the order that they were actually shutdown in was not
guarenteed. If the workers were shutdown first before the consumers.
Then the consumer could be deadlocked and have grabbed a job to process
but then have no worker to dispatch it to. This created a soft deadlock
in the system that would cause timeouts.

The new flow breaks consumers and workers into two groups of channels,
when the close signal is sent to the background processor the consumers
are shutdown first. And only once the consumers have entirely shut down
are the workers then terminated. This way if a job is consumed in the
same instant that the shutdown signal is sent, that job will still be
processed rather than being dropped or left in a bad state or causing a
deadlock.