feat(ui): Don't request all spending objects for transactions.
Only request the spending objects that are currently in view. This will
increase the overall number of requests but will make the requests
smaller and more efficient. It will also make it easier to support
soft-deleted spending objects in the near future. Which would be
excluded from the list endpoint, but can be requested directly via the
by ID endpoint.