fix(billing): Added a way to manage subscription from setup screen
If you had a subscription but somehow had no links setup, then there was
no way to manage the subscription or cancel it. You would have to setup
a link then cancel/manage the subscription.

Now there is a way to manage the subscription directly from the setup
page if necessary.

Resolves #1715