feat(ui): Adding more transaction upload UI functionality
The status for the transaction upload is now supplied to the client via
a web socket from the API. This websocket will deliver the entire
transaction upload object as the first message, and will provide changes
to the status of the upload in subsequent messages. If the upload is
already in a final state when the socket is opened, then only the
initial message will be sent then the socket will be closed.