fix(cron): Fixed bug that could cause duplicate cron jobs
I believe there was a rounding bug here where now could be just in the
right time where the next cron job would still be the same as the
current timestamp or before. Causing it to process twice. This adds a
bit of padding to the timestamp to make sure that we do push the cron
out to the next one even if the cron is processing every second.

I think

Resolves #1774