feat(api): Laying some minor groundwork about extending tokens
I want to be able to allow tokens to be automatically renewed for users
who frequent the application. This way if you are using the application
somewhat frequently, you will not be prompted to reauthenticate as
frequently.

Essentially a duration will be defined, this duration is the amount of
time before the expiration of a token that a token can be reissued if it
is observed by the API. When a token is reissused its reissuer counter
is incremented. Once a token reaches a certain number of reissues it
will no longer be reissued. If a token expires naturally before it can
be reissued then it will simply be expired.

More work for this to come.