fix(dev): Remove support for gitpod
I'm just not maintaining this at all, once upon a time it was something
I wanted to use for easier remote development when I travelled. But I've
used it probably once, and it currently does not work. Codespaces is
probably easier for more people to use, and seems to still work so I'm
going to remove gitpod and leave that in.

Resolves #1639