fix(config): Fixed how the external URL and protocol was determined
Previously monetr would rely on 3 fields to determine the external URL
and protocol used to access the application. This was confusing and was
inconsistent throughout the application.

This transitions monetr to using a single setting,
`$.server.externalUrl` which is a simple URL. monetr then parses this
URL and derives the protocol, hostname, and path prefix if there is one.
URLs for monetr are then built based on this base URL.

Resolves #1800