fix(storage): Improved path validation in filesystem storage provider
The filesystem storage provider now provides better validation for the
provided base path. Requiring that the path is an absolute path and that
the process has permissions to access the directory.

This also adds MONETR_STORAGE_PROVIDER to the cmake configuration,
allowing the storage provider to be customized between S3 and the
filesystem provider.

Resolves #1798