fix(forecast): Make the forecasting code return a proper error
This code previously would just panic if there was a problem, which
sucked because it did handle timeouts. So if it did timeout it would
simply panic which was not helpful.

This corrects all of the interfaces such that they now properly return
an error to the caller when there is a timeout or other problem.

Resolves #1793