feat(sentry): Fork sentryecho middleware
The sentry echo middleware recently added performance tracing in the
official sentry SDK. However their implementation of the tracing is
different from what I've already established in monetr. For the time
being I'm going to fork the middleware (located in
server/internal/sentryecho) and use that. Eventually I will move over to
the new middleware if there is a reason to.