feat(api)!: Moving from audience to scope
This refactors how tokens are treated in the API and its middleware. Now
the authentication token is parsed on almost every request, regardless
of whether or not an endpoint requires authentication. This will allow
us to attribute some requests to specific users (such as /api/config)
but it also allows us to dynamically scope specific endpoints behind
specific token scopes. For example, the /api/users/me endpoint should be
accessible if you are authenticated via MFA or fully authenticated. but
the rest of the endpoints should only be accessible if you are fully
authenticated.

This should give monetr a lot more flexibility when it comes to scoping
endpoints, and even lends itself to being able to allow admin endpoints
and tuck those behind a scope that regular users will not have.