feat(cmd): Adding admin command to migrate between KMS providers
monetr is moving away from GCP and instead moving to a co-located
datacenter. To that end, we need to migrate all of the data encrypted
using Google Cloud's KMS, to our new KMS system backed by Hashicorp
Vault.

This lays the groundwork of an admin function to do exactly that.