feat(backup): Adding support for backups.
This adds a backup utility to monetr that will take all of the data from
the database, and object store/filesystem. It will then write all of
that data to a single tar.gz file. This is intended to be done with as
little memory overhead as possible and allows the file to be streamed to
an object store.

This is a work in progress and is very much just a fun experiement. If
the design works though, the monetr will be able to create backups for
an entire production size dataset into a single file all with a minimal
memory overhead as well as limited disk space.

A restore utlity will also need to be created.