fix(test): Improved speed of detecting Go tests
All of monetr's tests are run via CTest, this requires pre-detecting
them and essentially building an array of test commands. One command for
each individual test. This involves executing a script called
`find_tests.go`, and previously this script was executed once per
package in the server/ directory. This changes it so that instead the
script is only executed once (and only when testing is enabled), and the
script then outputs JSON which is easier to parse in cmake and allows
for cleaner and easier to understand code.