fix(ui): Fixing uncontrolled path bug in UI router (#1853)
* fix(ui): Fixing uncontrolled path bug in UI router

Github's security thing is complaining about using a raw path from the
user's input to load files. But there really is no risk because the
filesystem that the UI is loaded from is embeded and is basically
sandboxed. There are no extra files that could possibly be requested
maliciously.

But I'm going to try to improve it anyway to keep the alert from comming
up again and again.

* chore: Cleaning up