fix(job): Fixed how remove link job found plaid transactions
It is possible for plaid transactions to become "orphaned", where there
is not a single transaction record that actually points to the plaid
transaction. In this scenario the transactions would not get removed and
the remove link job would fail because it would not be able to delete
bank accounts for that plaid link due to a foreign key constraint.

This resolves the issue by not trying to find plaid transactions using
their transaction association, but instead just that the plaid
transaction is related to a bank account.