fix(plaid): Improve transaction clustering for Plaid syncs
Because of how monetr and plaid handle transitions of pending to cleared
for transactions, the data around a transaction can sometimes get skewed
and have the wrong strings for transaction names and merchants because
of poor association between the pending transaction and the cleared
transaction.

This attempts to help resolve the clustering issue, by updating
transactions to have the names/merchants of the most recent data
received from plaid. So if we have a pending transaction and we receive
the cleared transaction data for the same item, we will overwrite the
name with the cleared name and merchant.

Resolves #1714