feat(dev): Adding local development "lite" environment (#2024)
The "lite" development environment allows developers to only work on the
frontend of monetr but still use a real API backend without needing to
run the entire stack locally. This development environment proxies API
requests upstream to a specified API server such as the staging
environment.

This allows for bugs to be reproduced more easily or for changes to be
made on systems that may not want or need to run all of the extra
servives necessary.