feat(comms): Notify users a few days before their trial expires
This adds support for sending an email to users a few days before their
trail expires. The way this is built, it should not send more than a
single email per user. It is possible that a user might not receive an
email if we mark the user as having this notification processed for
them, but then an upstream SMTP failure prevents it from properly sent.
But we should not be able to send the email multiple times.