feat(api): Add a role concept to users, specifically owner
In order to send trial notifications, we need to handle the possibility
that an account has more than one user. To that end, only a single user
is seen as "responsible for billing" in monetr's code. Therefor only
that "responsible" user should receive a notification for something like
a trial ending.

This adds a role to the user object (and removes an old billing column
that shouldn't be used anyway), and makes sure that the oldest user (by
ID) per account, is now considered the "owner" of the account. Then it
makes sure that all other users in the application are considered
members.