feat(comms): Notify users a few days before their trial expires (#2028)
* feat(comms): Notify users a few days before their trial expires

This adds support for sending an email to users a few days before their
trail expires. The way this is built, it should not send more than a
single email per user. It is possible that a user might not receive an
email if we mark the user as having this notification processed for
them, but then an upstream SMTP failure prevents it from properly sent.
But we should not be able to send the email multiple times.

* chore: Fix broken tests

* feat(api): Add a role concept to users, specifically owner

In order to send trial notifications, we need to handle the possibility
that an account has more than one user. To that end, only a single user
is seen as "responsible for billing" in monetr's code. Therefor only
that "responsible" user should receive a notification for something like
a trial ending.

This adds a role to the user object (and removes an old billing column
that shouldn't be used anyway), and makes sure that the oldest user (by
ID) per account, is now considered the "owner" of the account. Then it
makes sure that all other users in the application are considered
members.

* chore: Progress on notification

* chore: Adding tests for trial expiry notification

* chore: More tests

* chore: Rename migration

Resolves #2023