fix(build): Including notes about requiring gcc compiler
monetr uses cgo in order to talk to the host system's locales API. This
allows monetr to correctly work with multiple types of currency (in the
future) and store them appropriately based on their individual formats
(monetr does not use floating points).

This makes sure that gcc is present on the host system when building
monetr, but will not prevent the config from completing. It will display
a dev warning though if gcc is missing.

Resolves #2031