fix(forecast): Improve performance of forecasting
The timezone fix for the forecasting code had an unintended side effect
where the rule was being cloned for every single event that was
generated by the forecasting instructions. The clone is hacky and rather
slow. So doing it inside a huge for loop was not great.

This instead lifts the ruleset clone into the constructor of the
instructions, so that it is only cloned once per instruction; then
reused inside the projections.