feat(build): Moving storybook to webpack
I don't want to use rsbuild for storybook and rspack for the actual UI.
I'd rather have consistent tooling; and webpack and rspack are very
consistent between each other.

This moves everything back to plain old webpack 5, which should
hopefully also make it easier to upgrade to a newer version of
storybook.