feat(build): Moving storybook to webpack and interface to rsbuild (#2051)
* feat(build): Moving storybook to webpack

I don't want to use rsbuild for storybook and rspack for the actual UI.
I'd rather have consistent tooling; and webpack and rspack are very
consistent between each other.

This moves everything back to plain old webpack 5, which should
hopefully also make it easier to upgrade to a newer version of
storybook.

* feat(ui): Moving the UI over to using rsbuild instead of rspack

* feat: Getting everthign working

* Getting moved over to rsbuild

* chore: Removing unused dependencies

* chore: Tweaks

* chore: Update license list

So this isn't entirely complete yet, I would like to move storybook over to using rsbuild as well; but as part of that migration i think it just makes sense to move it to the latest version of storybook as well.

So assuming the build passes, im just going to merge this as is. It removes rspack from the storybook and moves interface over to rsbuild.