feat(math): Adding fourier caluclations and notes
I think that I could build a recurring transaction detection thing that
uses a fourier transform to find out if and how frequently a group of
transactions recurr.

This is an initial commit which contains a shitload of notes as I try to
understand the algorithm (which might not be implemented correctly). My
end goal with this is to implement a pure go version as well as an X86
assembly version(s) leveraging both AVX and AVX512 features to perform
the math.