fix(dev): Fixed local development on macOS
When developing locally on macOS, the sass embedded package was being
installed in an odd location which was preventing it from being resolved
properly for some reason. But making it so that is used by multiple
projects makes sure that it will always be in the top pnpm node modules
directory which fixes the problem.