feat(ui): Add next due date to expense list view
This adds a little badge to the list view next to the name of the
expense. The name will be truncated if it is too long, and the badge
will always be visible on both desktop and mobile views.

This also fixes a bug on the expense details page where hitting "Enter"
would prompt you to delete the expense instead of simply saving the
expense.

Resolves #1501