fix(ui): Indicate if a bank account is inactive
Inactive bank accounts are ones who are no longer showing up in the
Plaid API, and therefore will not be receiving updates. This could
happen for any number of reasons, one of the common ones being that the
account has simply been closed on the bank's side.

This adds a badge on the bank account dropdown that indicates if an
account is "inactive".

Resolves #1082