fix(ui): Fix pull to refresh on detail pages
The pull to refresh was taking into account scroll behavior from the
actual content of the pages. This makes it so that it won't anymore.

Eventually I need to just redo how the scroll containers work entirely
and instead have the entire page scrollable and have fixed and floating
UI elements instead so they are not affecting how scrolling works on
each page.