feat(ui): Revamping manual link setup flow
This is a sort of rewrite of the manual link setup flow. It breaks the
pieces into individual components as well as addresses some of the pain
points (at least) I had. For example; not being able to hit enter to go
to the next step. This should allow the onboarding flow to be a bit more
customizable in the long run as well.