feat(docs): Eject embeded UI from the documentation site
I had made it possible to use UI components directly in the
documentation site; however this brought a lot of complexity and
managing those components is a lot more two sided than I thought it
would be.

My goal was to put the component into the docs site once, and then never
have to touch the component again because as the component would change
and evolve. Styles would be tweaked, colors would change; the docs site
would always have a perfect up-to-date representation of the component
to show a user. But what I got was a massive bundle size, competing
stylesheets. Nesting react-router inside of next/router. And a shitload
of complexity.

I still really like this idea and I think there is a place for it; but I
dont think the home page is that place. Maybe embeded storybooks will
suite the documentation pages themselves well.

Come back to this if I ever need to embed the components again!