feat(build): Remove licensed from build process (#2273)
I had intended to use this to create a bill of materials for the entire
application as well as some basic license compliance testing. However it
has not served its purpose well and adds a significant amount of time to
the build process. And now for some reason has broken entirely on Debian
12.

The package is no longer terribly active and isn't receiving updates
besides basic dependendency changes.

I think I'll need to find something better in the long run.