fix(ui): Fixed type error in selected bank account hook
Somehow the initial data that this hook is derived from would be a
non-undefined thing, like an object? But that would cause this hook to
fail because it's expecting an array. So I've added a type check to only
derive the initial data from the other query state _if_ that query state
is an array.