fix(docs): Improve documentation around external URL
monetr uses the external URL configuration as a source of truth for
setting the cookie for authentication, sending links in emails, and
more. Because of this it can be misleading when monetr still loads valid
web pages on another address but then fails to authenticate.

I've improved the documentation around this to try to help aleviate any
confusion, but ultimately more improvements need to be made as feedback
is received.

Related to #2287