feat(manual): Adding support for additional manual accounts on one link (#2299)
* feat(manual): Adding support for additional manual accounts on one link

This is adding support for people to create additional manually
budgeted bank accounts on the same manual link.

* feat(ui): Adding modal for creating additional manual accounts

Adds a modal that allows users to create additional bank accounts for
manual links. The modal does not restrict itself to only manual links
though, but the API will return an error if you try to create an account
on a Plaid managed link.

* chore: Addressing deepsource warnings

* chore: Improvements

* fix failing test