feat(billing): Allow users to subscribe before their trial ends
If someone wants to subscribe to monetr before their trial has ended
they will now be able to do so. This will not trample the trial data,
but will make the subscription status take priority. It also makes sure
that we don't send trial expiry notifications when we don't need to.