fix(locale): Support for non-default currency accounts (#2318)
* feat(api): Add support for non-default currency accounts

Now when creating a manual link/account it will use your locale's
default currency. Transactions created on this account will use the
currency of the account unless they are created via file upload (might
be problematic later?).

* feat: Support user's default currency

fixlocale): Complete rewrite of how the frontend formats currency

This is a complete change in how the frontend formats currency to and
from the stored amounts. This will make it possible for the frontend to
seamlessly support other currencies.

* chore: Cleanup

* chore: Fixing tests

fix(ofx): Make sure the currency of the file matches the account

If we upload an OFX file that has a currency code different from the
bank account the file is being uploaded into; then fail as this file
could corrupt account data.

* chore: Fuck floating point numbers

* fuck floating point numbers

* chore; Adding test