chore(compose): Updating compose files valkey version
We really don't need a precise version of valkey as it is not a huge
load bearing part of monetr.

I've also added a note to the compose directory explaining that it
should not be used as a reference for self-hosted instances, the compose
directory is for local development only.