feat(ui): Allow currency of bank account to be changed
This is the last piece needed for someone to actually customize the
currency of a manual bank account in monetr.

This will also include some quality of life changes around the bank
account picker, including moving the "Add account" button into the
actual bank account list which seems to make more sense.

Resolves #2330