fix(ofx): Allow an alternative OFX date format to be parsed
User has a slightly different date format in their OFX file, monetr was
being too strict about the format so it wasn't parsing it. This makes it
a bit more flexible. But if there are more date formats that I haven't
seen yet then there will still be bug reports in the future.

Resolves #2362