fix(api): Fix not handling long passwords
If a password is provided that is longer than 72 bytes then bcrypt wont
be able to handle it. This might be another indicator to start looking
into argon2 instead, but idk what it's limitations are.

In the mean time, return a proper error for this situation as well as
validate the input on the frontend.

Resolves #2367