feat(api): Improving transaction similarity API experience (#2389)
* feat(api): Improving transaction similarity API experience

This work definitly hurts the efficiency of the transaction cluster
calculation, I think I can improve some of that later on. But in the
mean time this should provide a really good starting point for providing
consistent identifiers for transaction clusters. Specifically, the
cluster calculation now generates a signature based on the top 5 most
valuable words in the cluster that have a value greater than 0. This way
words that push the clusters closer together will end up making up the
signature. The most valuable words of those should be consistent. I
might need to reduce the number of words to consider down to something
like 2 though.

* chore: Improvements