fix(ui): Use account timezone if available instead of browser timezone
If a user has their timezone configured as X and their browser timezone
is Y. This can cause some of the dates sent to the server to be
malformed and can cause some odd behaviors. For lack of a better
solution for now (like not having times at all and instead only using
dates?); this adds timezone to components that need it. Forcing the time
to be shown and calculated in the user's timezone instead of only the
browsers timezone.

Resolves #2386