feat(ui): Allow currency of bank account to be changed (#2358)
* feat(ui): Allow currency of bank account to be changed

This is the last piece needed for someone to actually customize the
currency of a manual bank account in monetr.

This will also include some quality of life changes around the bank
account picker, including moving the "Add account" button into the
actual bank account list which seems to make more sense.

Resolves #2330

* fix: Progress on getting the settings UI ready

* chore: Adding bank account settings page

* chore; Adding tests to use x/text/currency

* chore: Tidy

* chore: Progress on bank account settings UI

* chore: Progress on currency changing

* chore: More improvements