fix(ui): Fix monetr logo rendering issues
Sometimes the monetr logo would not render properly when the logo
appeared on the screen multiple times. This was due to how the gradients
were being applied by ID in the SVG. This changes it so the SVG is
imported and rendered as an img tag, preventing this problem going
forward.