feat(locale): Migrating to CLDR as a locale datasource
Right now monetr uses glibc's `locales.h` as a datasource for numeric
currency data and locale information. This dataset is not really as up
to date as I had hoped and does not contain as much information as I've
found I need.

CLDR however is much more extensive and is used by a lot of other locale
packages. So it should be more suitable.