feat(api): Adding PATCH functionality to API
This change is an effort to support more complex API calls with ease via
patch requests to the API. Golang is not terribly good at this so it
requires a substantial amount of boiler plate. But this change should
make it easy to create more PATCH endpoints going forward as well as
abstract a lot of the logic away.