fix(plaid): Fix Plaid pagination error handling (#2454)
When we get an error from Plaid indicating that the sync must be retried
from scratch we now throw an error causing the transaction to be rolled
back. We then retry the job from scratch up to a few attempts to correct
the issue. If we get any other error then we still return it normally.

Resolves #2453