docs: Document getting monetr running on Kubernetes (#2439)
* feat(docs): Document getting monetr running on Kubernetes

This is a basic guide on how to get monetr running in Kubernetes. It
assumes that the user is familiar with the tooling available and
understands how to run Kubernetes and its nuances.

This guide will **not** be a guide on how to install or manage monetr's
dependencies within Kubernetes, such as Valkey, PostgreSQL or storage
systems.

However the guide does include recommendations for such things based on
monetr's own production deployment. But it is up to the end user to
learn and manage those tools on their own.

* feat(docs): Adding guides and roadmap

* chore: tWeaks, moving roadmap out of this PR

* chore: Fix trailing slash

* chore: Cleanup

* chore: Tweaks

* chore: Fixing