fix(api): Fixed transaction clusters not having a name
It was possible for transaction clusters to be calculated without a
name, this should be considered a bug but until I've had more time to
investigate it I'm just going to omit clusters without a name from being
persisted to the database.

What is likely happening is this "no name" cluster is a bunch of
transactions that are so close to the 0 point, or consisting of such low
value words in their description that there aren't any non-zero words to
use for a name.

Resolves #2475