fix(api): Significantly improve logging
This adds a shitload of logging to a ton of different areas, including
logging changes to budgets as they are calculated to help better
diagnose issues in the future. The logs produce enough information that
if a bug is reported in the calculations I should be able to compose a
test to reproduce the bug with the informataion.

This also changes several call signatures of the code internally in
order to more easily pass data through, like passing actual timezone
objects instead of just the timezone string.